Ban Users

+++ back_end/src/main/java/com/blog_01/controller/AdminController.java
@@ -1,11 +1,14 @@
 package com.blog_01.controller;
 
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -32,11 +35,14 @@ public class AdminController {
         }
 
         String token = authHeader.substring(7);
-        Long id = jwtService.extractId(token);
+        String role = jwtService.extractRole(token);
+        if (!role.equals("ADMIN")) {
+            return ResponseEntity.status(401).build();
+        }
         return ResponseEntity.ok(adminService.getAllUsers());
     }
 
-     @DeleteMapping("delete/{postId}")
+    @DeleteMapping("/delete/{postId}")
     public ResponseEntity<?> deletePost(
             @PathVariable Long postId,
             @RequestHeader("Authorization") String authHeader
@@ -45,12 +51,33 @@ public class AdminController {
             return ResponseEntity.status(401).build();
         }
         String token = authHeader.substring(7);
-        Long id = jwtService.extractId(token);
+        String role = jwtService.extractRole(token);
+        if (!role.equals("ADMIN")) {
+            return ResponseEntity.status(401).build();
+        }
         boolean isDeleted = adminService.deletePost(postId);
         if (isDeleted) {
             return ResponseEntity.ok().build();
         } 
         return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        
+    }
+
+    @PostMapping("/ban/{userId}")
+    public ResponseEntity<?> banUser(
+            @PathVariable Long userId,
+            @RequestHeader("Authorization") String authHeader
+    ) {
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            return ResponseEntity.status(401).build();
+        }
+        String token = authHeader.substring(7);
+        String role = jwtService.extractRole(token);
+        if (!role.equals("ADMIN")) {
+            return ResponseEntity.status(401).build();
+        }
+        adminService.banUser(userId);
+        return ResponseEntity.ok(Map.of("message", "Utilisateur banni/débanni avec succès"));
         
     }
 }
